// src/FX.d.ts
declare enum ComponentType {
    Sprite = 0,
    MovieClip = 1
}
// ...
export interface IBaseEffect {
    name: string;
    id: any;
    type: number;
    containerId: string;
}
export interface IEmitterSettings extends IBaseEffect {
    core: ICoreSettings;
    spawnFrequencyMin: number;
    spawnFrequencyMax: number;
    particleSettings: IParticleSettings;
    maxParticles: number;
    spawnCountMin: number;
    spawnCountMax: number;
    duration: number;
    infinite: boolean;
    useGravity: boolean;
    gravity: number;
    useFloor: boolean;
    floorY: number;
    rotation: number;
    autoRotation: number;
    childs: IEmitterSpawn[];
}
export interface IEmitterSpawn {
    id: string;
    type: number;
    scale: number;
    name: string;
    adoptRotation: boolean;
    containerId: string;
}
export interface IEmitterSpawns {
    onStart: IEmitterSpawn[];
    onHalfway: IEmitterSpawn[];
    onBounce: IEmitterSpawn[];
    onComplete: IEmitterSpawn[];
}
// ...
export interface ISpriteSettings {
    texture: string;
    anchorX: number;
    anchorY: number;
}
export interface IMovieClipSettings {
    textures: string[];
    anchorX: number;
    anchorY: number;
}
export interface ICoreSettings {
    type: string;
    params: ICircleCoreParams | IRingCoreParams | IBoxCoreParams;
}
export interface ICircleCoreParams {
    radius: number;
    radial: boolean;
    angle: number;
}
export interface IRingCoreParams {
    radius: number;
    radial: boolean;
    angle: number;
    uniform: boolean;
}
export interface IBoxCoreParams {
    width: number;
    height: number;
    radial: boolean;
}
export interface IParticleSettings {
    componentType: ComponentType;
    componentId: string;
    componentParams: IBaseComponentParams;
    durationMin: number;
    durationMax: number;
    useMotion: boolean;
    useRotation: boolean;
    useAlpha: boolean;
    useScale: boolean;
    useTint: boolean;
    useChilds: boolean;
    useSpawns: boolean;
    distanceMin: number;
    distanceMax: number;
    distanceEase: string;
    moveSpeedMin: number;
    moveSpeedMax: number;
    bounceFacMin: number;
    bounceFacMax: number;
    frictionMin: number;
    frictionMax: number;
    align: boolean;
    blendMode: number;
    addOnTop: boolean;
    rotationSpeedMin: number;
    rotationSpeedMax: number;
    randomRotationDirection: boolean;
    randomStartRotation: boolean;
    fadeIn: boolean;
    fadeInDurationFac: number;
    fadeInEase: string;
    alphaStartMin: number;
    alphaStartMax: number;
    alphaEndMin: number;
    alphaEndMax: number;
    alphaEase: string;
    tintStart: number;
    tintEnd: number;
    tintEase: string;
    scaleIn: boolean;
    scaleInDurationFac: number;
    scaleInEase: string;
    uniformScale: boolean;
    scaleXStartMin: number;
    scaleXStartMax: number;
    scaleXEndMin: number;
    scaleXEndMax: number;
    scaleXEase: string;
    scaleYStartMin: number;
    scaleYStartMax: number;
    scaleYEndMin: number;
    scaleYEndMax: number;
    scaleYEase: string;
    scaleStartMin: number;
    scaleStartMax: number;
    scaleEndMin: number;
    scaleEndMax: number;
    scaleEase: string;
    stopOnBounce: boolean;
    spawn: IEmitterSpawns;
    childs: IEmitterSpawn[];
}
export interface IMovieClipComponentParams extends IBaseComponentParams {
    animationSpeedMin: number;
    animationSpeedMax: number;
    loop: boolean;
}
export interface IBaseComponentParams {
    anchorX: number;
    anchorY: number;
}